parser: used unsigned where appropriate.
	* src/ast/factory.hh, src/ast/factory.cc
	(make_external_event_or_function): Take arity as an unsigned.
	* src/parser/ugrammar.y: Adjust.

+++ src/ast/factory.cc
@@ -626,7 +626,7 @@ namespace ast
   rExp
   Factory::make_external_event_or_function(const location& l,
                                            libport::Symbol kind,
-                                           float arity,
+                                           unsigned int arity,
                                            libport::Symbol obj,
                                            libport::Symbol slot,
                                            libport::Symbol id) /* const */
